✨ feat: AI 추천 API(JSON 응답) 추가
